silence when ticks stop. (and long is 32bit on 32bit platform

+++ switch.h
@@ -4,7 +4,7 @@ struct model {
   int pst_b;
   int fader;
   int start_fader;
-  long tick;
+  long long tick;
 };
 
 #ifdef __cplusplus
